Some major bug fixes

+++ plans/views.py
@@ -1,43 +1,51 @@
 from typing import Optional
-from rest_framework.response import Response
-from rest_framework.exceptions import NotFound
-from rest_framework.request import Request
-from rest_framework.decorators import api_view, permission_classes
-from rest_framework.permissions import IsAuthenticated
 from rest_framework import status
+from rest_framework.request import Request
+from rest_framework.response import Response
+from rest_framework.exceptions import NotFound,  APIException
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.decorators import api_view, permission_classes, parser_classes
+from rest_framework.parsers import MultiPartParser, FormParser
+from django.http.request import QueryDict
 
-from .serializers import BudgetSerializer, GoalSerializers, ReminderSerializer
-from .models import Budget, Goal, Reminder
+from .models import Budget, Goal
+from .serializers import BudgetSerializer, GoalSerializers
 
 
-@api_view(http_method_names=['GET', 'POST'])
+@api_view(http_method_names=['GET'])
 @permission_classes([IsAuthenticated])
-def goals(request: Request) -> Response:
+def goals(request: Request) -> Optional[Response]:
+    goals = Goal.objects.filter(user=request.user)
+    serialized_goals = GoalSerializers(goals, many=True)
+    return Response(serialized_goals.data, status=status.HTTP_200_OK)
 
-    if request.method == 'GET':
-        goals = Goal.objects.filter(user=request.user)
-        serialized_goals = GoalSerializers(goals, many=True)
-        return Response(serialized_goals.data, status=status.HTTP_200_OK)
 
-    if request.method == 'POST':
-        data = request.data.copy()
-        data['user'] = request.user.pk
-        serialized_goals = GoalSerializers(data=data)
-        if serialized_goals.is_valid():
-            return Response(serialized_goals.data, status=status.HTTP_201_CREATED)
-        return Response(serialized_goals.errors, status=status.HTTP_400_BAD_REQUEST)
+@api_view(http_method_names=['POST'])
+@parser_classes([MultiPartParser, FormParser])
+@permission_classes([IsAuthenticated])
+def create_goal(request: Request) -> Response:
+    data: QueryDict = request.data.copy()
+    data['user'] = request.user.pk
+
+    serialized_goals = GoalSerializers(data=data)
+    if serialized_goals.is_valid():
+        serialized_goals.save()
+        return Response(serialized_goals.data, status=status.HTTP_201_CREATED)
+    return Response(serialized_goals.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 @api_view(http_method_names=['PUT', 'DELETE'])
 @permission_classes([IsAuthenticated])
+@parser_classes([MultiPartParser, FormParser])
 def upgrade_goals(request: Request, pk: int) -> Response:
 
     if request.method == 'PUT':
         current_goal = Goal.objects.filter(pk=pk, user=request.user).first()
         data = request.data.copy()
-        data['user'] = request.user.pk
+        data["user"]= request.user.pk
         goal_serializer = GoalSerializers(current_goal, data=data)
         if goal_serializer.is_valid():
+            goal_serializer.save()
             return Response(goal_serializer.data, status=status.HTTP_202_ACCEPTED)
         return Response(goal_serializer.errors, status=status.HTTP_417_EXPECTATION_FAILED)
 
@@ -45,26 +53,54 @@ def upgrade_goals(request: Request, pk: int) -> Response:
         current_goal = Goal.objects.filter(pk=pk, user=request.user).first()
         if current_goal:
             current_goal.delete()
-            return Response({'message': 'successfully removed'}, status=status.HTTP_204_NO_CONTENT)
+            return Response({'detail': 'successfully removed'}, status=status.HTTP_204_NO_CONTENT)
         raise NotFound(detail="Goal don't extsts")
 
 
-@api_view(http_method_names=['DELETE'])
+@api_view(http_method_names=['PUT', 'DELETE'])
 @permission_classes([IsAuthenticated])
-def remove_budget(request: Request, pk: int) -> Response:
+def change_budget(request: Request, pk: int) -> Response:
+
+    if request.method == 'PUT':
+        current_budget: Budget = Budget.objects.filter(
+            pk=pk, user=request.user).first()
+
+        if current_budget.has_expired:
+            raise APIException('An expired budget cannot be deleted',
+                               code=status.HTTP_417_EXPECTATION_FAILED)
+
+        data: dict = {**request.data, 'user': request.user.pk}
+
+        if data.get('total_amount') <= sum([exp.amount for exp in current_budget.expenses_set.all()]):
+            raise APIException('This amount is already is used in some expense ',
+                               code=status.HTTP_424_FAILED_DEPENDENCY)
+
+        budget_serializer = BudgetSerializer(current_budget, data=data)
+        if budget_serializer.is_valid():
+            budget_serializer.save()
+            return Response(budget_serializer.data, status=status.HTTP_202_ACCEPTED)
+        return Response(budget_serializer.errors, status=status.HTTP_417_EXPECTATION_FAILED)
+
+    # direct removal of budget are allowed if the expenses relation.length is zero
 
     if request.method == 'DELETE':
-        budget_exists = Budget.objects.filter(pk=pk).first()
+        budget_exists = Budget.objects.filter(pk=pk, user=request.user).first()
 
         if budget_exists:
-            budget_exists.delete()
-            return Response({'message': 'Budget has been deleted successfully'}, status=status.HTTP_302_FOUND)
+            if budget_exists.expenses_set.count() == 0 or budget_exists.has_expired:
 
-        raise NotFound(detail="Budget don't exists")
+                budget_exists.delete()
+                return Response({'detail': 'Budget has been deleted successfully'}, status=status.HTTP_202_ACCEPTED)
+
+            raise APIException(
+                'An un-expired budget cannot be removed remove all the expenses assocaited with it first', code=status.HTTP_406_NOT_ACCEPTABLE)
+
+        raise NotFound(detail="Budget don't exists ",
+                       code=status.HTTP_404_NOT_FOUND)
 
 
-@api_view(http_method_names=['GET', 'POST'])
-@permission_classes([IsAuthenticated])
+@ api_view(http_method_names=['GET', 'POST'])
+@ permission_classes([IsAuthenticated])
 def budget(request: Request) -> Response:
 
     if request.method == 'GET':
@@ -82,37 +118,6 @@ def budget(request: Request) -> Response:
             serialized_budget.save()
             return Response(serialized_budget.data, status=status.HTTP_201_CREATED)
         return Response(serialized_budget.errors, status=status.HTTP_424_FAILED_DEPENDENCY)
-
-
-@api_view(http_method_names=['GET', 'POST'])
-@permission_classes([IsAuthenticated])
-def reminders(request: Request) -> Response:
-
-    if request.method == 'GET':
-        reminders = Reminder.objects.filter(user=request.user)
-        reminder_serializer = ReminderSerializer(reminders, many=True)
-        return Response(reminder_serializer.data, status=status.HTTP_200_OK)
-    if request.method == 'POST':
-        data = request.data.copy()
-        data['user'] = request.user.pk
-
-        reminder_serializer = ReminderSerializer(data=data)
-        if reminder_serializer.is_valid():
-            reminder_serializer.save()
-            return Response(reminder_serializer.data, status=status.HTTP_201_CREATED)
-        return Response(reminder_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-
-@api_view(http_method_names=['DELETE'])
-@permission_classes([IsAuthenticated])
-def delete_reminders(request: Request, pk: int) -> Optional[Response]:
-
-    if request.method == 'DELETE':
-        # reminder = Reminder.objects.filter(pk=pk,user=request.user).first()
-        reminder = None
-        if reminder:
-            return Response({'detail': 'Removed reminder'}, status=status.HTTP_302_FOUND)
-        raise NotFound(detail="Invalid id")
 
 
 # @api_view(http_method_names=['GET'])
